+++ src/ui/display.h
@@ -9,6 +9,7 @@ extern char g_gas_fee[30];
 extern char g_struct[120];
 extern char g_function[120];
 extern char g_amount[30];
+extern int g_is_token_listed;
 
 /**
  * Display address on the device and ask confirmation to export.
@@ -39,3 +40,6 @@ int ui_prepare_tx_aptos_account_transfer(void);
 
 int ui_display_tx_coin_transfer(void);
 int ui_prepare_tx_coin_transfer(void);
+
+int ui_display_tx_fungible_asset_transfer(void);
+int ui_prepare_tx_fungible_asset_transfer(void);