+++ CHANGELOG.md
@@ -5,11 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.8.13]
+
+### Changed
+
+- Re-added the ability to sign cointypes not listed in the app. Same capacity added to the Fungible Assets.
+
+## [0.8.12]
+
+### Added
+
+- Add support for fungible assets trading
+
+### Changed
+
+- Changed screens tor coin_transfer to support Clear Signing Guidelines
+
+## [0.7.11]
 
 ### Added
 
 - Add support for Flex target.
+- Add support for Ledger Swap
 
 ### Changed